Fix ELF offset

+++ src/util/util.h
@@ -40,7 +40,7 @@ char *join_paths(const char *paths[]);
 #define JOIN_PATHS(...)  join_paths((const char*[]){__VA_ARGS__, NULL})
 char *get_ext(const char *filename);
 char *change_ext(const char *path, const char *ext);
-void put_padding(FILE *fp, uintptr_t start);
+void put_padding(FILE *fp, long start);
 bool is_file(const char *path);
 
 void show_version(const char *exe);
